Refactoring to support multiple pages, breaking things!
Refactored to break things into their logical "pages" using react-router. Completely removed the Player component and merged it into the top level App. Set up grabbing the player from a fetch() of a stub to simulate a backend. Splitting the app into multiple pages causes the app's state to be reset everytime a new route is followed, so the next step is to figure out a way around that...which might mean building a backend app :s