learned Pixi.js, got a demo app working...kinda
Build the demo app for Pixi.js and then started messing with it to get it talking to the React player management app. Not sure on my architecture here yet...this might be where a server needs to exist to direct traffic? For now using localstorage to set variables with the game that are then consumed by the react app.