GOT THE GAME FRAMEWORK WORKING!!!
Came up with a way to render the pixi.js game inside a react component instead of doing the weird bouncing back and forth I was doing. This should now be a good template for all of the micro games, as I plan to generate a list of games to play and pass them to the game component to instruct the pixi app which scenes to load. This should handle the whole chain from game to game, and then kick the player back out to the player UI when the game loop is complete. Wooo!!!!