fixed game switching, added cash scoring, refacroted to move each game into own module, fixed bug in treaurehunter
The system is now creating a mix of games and swaping between them based on each game's logic and recording the correct amound of cash earned if the player succeeds. Moved both games' setup and play functions to seperate files, probably using too many this statements, but since I had to bind it anyway, I was lazy and didn't remove the rest. I'll refoctor or jsut do better when I build the real microgames next. Also fixed the bug in TreasureHunter that was causing all the blobs to line up in a nice neat line like jerks.