BROKEN!! but not as broken
Soo....ejecting didn't work. It broke A LOT of other stuff. I had to create a new create-react-app and move stuff over, so I'm now back to where I was at the beginning of all of those, unable to get the totally not modular Phaser-CE libary to load using create-react-app's webpack setup. I may have to bite the bullet and live on the bleeding edge with Phaser3, which doesn't use PIXI but is modular...