FIXED!
Our long national nightmare is over. Eventually I will need to actually eject this app and go through all the work that is conifguring (and removing) all the shit it's currently doing for me...but that's not the point of this project, the point of this project is to build a react app that is also a game. So onward and upward, I'm not sure what kind of weird ESLint config is being used under the covers, and since I can't just add to the existing webpack config, using eslint-disable comments to get around the boiler plate was the fastest way to fid things, even though it wasn't the best way.