Player UI 90% finished
Lots of pushing pixels around, made a placeholder sprite to see if my planned art style meshed with the UI style, added the logic to only show the mount sprite when a mount is equipped, added the tabs for equipped items between player and mount items, tweaked the UI a bit for smaller laptop screens. poshised the graphic a bit.