Player sprite responds to items being equipped, Mount sprite only shows when a mount is equipped
Lots of pixel art work making today to test this. All the code is in place for the player items to show up on the sprite in the player screen when an item is equipped now. Still need to do this for the mount and the mount items. Also added a spotlight selected image that I haven't hooked up yet to show which sprite is "active"