More experiments in animating the generation of lootboxes
I created what might be the worst looking treasure chest sprite in the history of drawing to allow me to play with the CSS on this. Still not finished but I think I have the way forward figured out.  TODOs in the code, hiopefully this will be a quick bit to finish up tomorrow.